Access go dependencies via existing debian packages
As our CI is building in an environment where access to external
services like github is not allowed we have to build with what we
get through the Ubuntu package archive. The packages we're using
here are used by the snapd debian package too so we're not basing
off something which isn't maintained.

This change will be also proposed upstream to have this available
as a general feature in snapcraft.