cmd: client: error out when calling user is not root
With the unix domain socket being owned by root we have another
security boundary in place, which is that only root can talk
to the service. To improve our user experience we should give
our users a hint what they need to do instead of simply failing
with a error message.